backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12

# Google Cloud
google-genai==1.0.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.0
google-cloud-logging==3.11.0
google-cloud-secret-manager==2.21.0

# Security
slowapi==0.1.9

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
